requirements.txt
tensorflow==2.15.0
streamlit==1.52.2
pillow>=9.0.0
pandas>=1.5.0
numpy>=1.23,<2
